refactor: useMonthDates 훅과 무관한 함수 분리 #36
